Graphics.SVGFonts.CharReference

Documentation

charsFromFullName :: String -> String

Parsing of xml character references.

I.e. "&#x2e;&#x2e;&#x2e;" is converted into a list of three Chars.

"ffb" is also parsed and converted into three Chars (not changing it).

characterStrings :: String -> [String] -> [Text]

A string represents a glyph, i.e. the ligature "ffi" is a string that represents the ligature glyph ffi