Graphics.SVGFonts.Fonts

Built-in fonts

bit :: (Read n, RealFloat n) => PreparedFont n

Bitstream, a standard monospaced font (used in gedit)

lin :: (Read n, RealFloat n) => PreparedFont n

Linux Libertine, for non-monospaced text. http://www.linuxlibertine.org/ Contains a lot of unicode characters.

lin2 :: (Read n, RealFloat n) => PreparedFont n

Linux Libertine, cut to contain only the most common characters. This results in a smaller file and hence a quicker load time.