Graphics.SVGFonts.ReadFont

Documentation

data FontData n

This type contains everything that a typical SVG font file produced by fontforge contains.

Constructors

FontData

Fields

fontDataGlyphs :: SvgGlyphs n

fontDataKerning :: Kern n

fontDataBoundingBox :: [n]

fontDataFileName :: String

fontDataUnderlinePos :: n

fontDataUnderlineThickness :: n

fontDataOverlinePos :: Maybe n

fontDataOverlineThickness :: Maybe n

fontDataStrikethroughPos :: Maybe n

fontDataStrikethroughThickness :: Maybe n

fontDataHorizontalAdvance :: n

fontDataFamily :: String

fontDataStyle :: String

fontDataWeight :: String

fontDataVariant :: String

fontDataStretch :: String

fontDataSize :: Maybe String

fontDataUnitsPerEm :: n

fontDataPanose :: String

fontDataSlope :: Maybe n

fontDataAscent :: n

fontDataDescent :: n

fontDataXHeight :: n

fontDataCapHeight :: n

fontDataAccentHeight :: Maybe n

fontDataWidths :: Maybe String

fontDataHorizontalStem :: Maybe n

This data is not available in some fonts (e.g. Source Code Pro)

fontDataVerticalStem :: Maybe n

This data is not available in some fonts (e.g. Source Code Pro)

fontDataUnicodeRange :: String

fontDataRawKernings :: [(String, [String], [String], [String], [String])]

fontDataIdeographicBaseline :: Maybe n

fontDataAlphabeticBaseline :: Maybe n

fontDataMathematicalBaseline :: Maybe n

fontDataHangingBaseline :: Maybe n

fontDataVIdeographicBaseline :: Maybe n

fontDataVAlphabeticBaseline :: Maybe n

fontDataVMathematicalBaseline :: Maybe n

fontDataVHangingBaseline :: Maybe n

bbox_dy :: RealFloat n => FontData n -> n

Difference between highest and lowest y-value of bounding box

bbox_lx :: RealFloat n => FontData n -> n

Lowest x-value of bounding box

bbox_ly :: RealFloat n => FontData n -> n

Lowest y-value of bounding box

underlinePosition :: RealFloat n => FontData n -> n

Position of the underline bar

underlineThickness :: RealFloat n => FontData n -> n

Thickness of the underline bar

horizontalAdvance :: RealFloat n => String -> FontData n -> n

Horizontal advance of a character consisting of its width and spacing, extracted out of the font data

kernAdvance :: RealFloat n => String -> String -> Kern n -> Bool -> n

Change the horizontal advance of two consective chars (kerning)

type OutlineMap n = Map String (Path V2 n)

A map of unicode characters to outline paths.

type PreparedFont n = (FontData n, OutlineMap n)

A font including its outline map.

loadFont :: (Read n, RealFloat n) => FilePath -> IO (PreparedFont n)

Read font data from font file, and compute its outline map.